完成了电影部分的 RecyclerView 视图，实现下拉刷新，向下加载更多的功能

+++ app/src/main/java/com/kyletung/doubanbookmovie/movie/MovieItemData.java
@@ -5,7 +5,7 @@ public class MovieItemData {
     private String movieId;
     private String movieName;
     private String movieImage;
-    private int moviePoints;
+    private double moviePoints;
     private String movieDirector;
     private String movieCasts;
     private String movieYear;
@@ -35,11 +35,11 @@ public class MovieItemData {
         this.movieImage = movieImage;
     }
 
-    public int getMoviePoints() {
+    public double getMoviePoints() {
         return moviePoints;
     }
 
-    public void setMoviePoints(int moviePoints) {
+    public void setMoviePoints(double moviePoints) {
         this.moviePoints = moviePoints;
     }
